SAM-6D/requirements.txt
# name: sam6d
# channels:
#   - xformers
#   - conda-forge
#   - pytorch
#   - nvidia
#   - defaults
# dependencies:
#   - pip=25.0.1
#   - python=3.12.10
#   - pip:

# These cannot be installed through requirements.txt because they need accompanying --index-url flag
#torch==2.3.1 # be careful when updating this version, torchvision and xformers need to also match, including CUDA support
#torchvision==0.18.1
#xformers==0.0.27

fvcore==0.1.5.post20221221
torchmetrics==0.10.3
blenderproc==2.8.0
opencv-python==4.11.0.86
# ISM
omegaconf==2.3.0
ruamel.yaml==0.18.10
hydra-colorlog==1.2.0
hydra-core==1.3.2
gdown==5.2.0
pandas==2.2.3
imageio==2.37.0
pyrender==0.1.45
pytorch-lightning==1.8.6 # previous versions are not compatible with pip version >23.x
pycocotools==2.0.8
distinctipy==1.3.4
git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf # SAM
ultralytics==8.3.113  # FastSAM
# PEM
timm==1.0.15
gorilla-core==0.2.7.8
trimesh==4.4.1
gpustat==1.0.0
imgaug==0.4.0
einops==0.8.1
